verification user author post and comment

+++ pkg/middleware/comment.go
@@ -29,8 +29,8 @@ func Updatecomment(db *sql.DB, comment *models.Comment) error {
 	return nil
 }
 
-func Removecomment(db *sql.DB, idcomment int64) error {
-	_, err := db.Exec("DELETE FROM comment WHERE id = ?", idcomment)
+func Removecomment(db *sql.DB, idcomment, currentuser int64) error {
+	_, err := db.Exec("DELETE FROM comment WHERE id = ? AND authorid = ? ", idcomment, currentuser)
 	if err != nil {
 		fmt.Println("Remove comment : ", err)
 		return err
